LIKE THIS, BIATCH

+++ ft_substr.c
@@ -6,23 +6,23 @@
 /*   By: fzaazaa <firas.zaazaa@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/12 17:36:33 by fzaazaa           #+#    #+#             */
-/*   Updated: 2024/12/15 11:52:26 by fzaazaa          ###   ########.fr       */
+/*   Updated: 2024/12/15 15:23:15 by fzaazaa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include"libft.h"
+#include "libft.h"
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	int		i;
-	int		true_len;
+	size_t	i;
+	size_t	true_len;
 	char	*sub;
 
 	i = 0;
 	true_len = 0;
-	if (!s || start >= (unsigned int)ft_strlen(s))
+	if (!s || start >= ft_strlen(s))
 		return (ft_strdup(""));
-	while (s[start + true_len] && true_len < (int)len)
+	while (s[start + true_len] && true_len < len)
 		true_len++;
 	sub = (char *)malloc(sizeof(char) * (true_len + 1));
 	if (!sub)
